Refactor thread action menus and dialogs
