+++ app/components/ThreadDeleteDialog.tsx
@@ -1,0 +1,55 @@
+import { Dialog, Button, Portal } from "@chakra-ui/react";
+
+interface ThreadDeleteDialogProps {
+  threadName: string;
+  onConfirm: () => void;
+  isOpen: boolean;
+  onOpenChange: (open: boolean) => void;
+}
+
+function ThreadDeleteDialog(props: ThreadDeleteDialogProps) {
+  const { threadName, onConfirm, isOpen, onOpenChange } = props;
+
+  return (
+    <Dialog.Root
+      role="alertdialog"
+      open={isOpen}
+      onOpenChange={({ open }) => onOpenChange(open)}
+    >
+      <Portal>
+        <Dialog.Backdrop />
+        <Dialog.Positioner>
+          <Dialog.Content>
+            <Dialog.Header>
+              <Dialog.Title>Are you sure?</Dialog.Title>
+            </Dialog.Header>
+            <Dialog.Body>
+              <p>
+                This action cannot be undone. This will permanently delete the
+                conversation <strong>{threadName}</strong> from our systems.
+              </p>
+            </Dialog.Body>
+            <Dialog.Footer>
+              <Dialog.ActionTrigger asChild>
+                <Button variant="outline">Cancel</Button>
+              </Dialog.ActionTrigger>
+              <Dialog.ActionTrigger asChild>
+                <Button
+                  colorPalette="red"
+                  onClick={() => {
+                    onOpenChange(false);
+                    onConfirm();
+                  }}
+                >
+                  Delete
+                </Button>
+              </Dialog.ActionTrigger>
+            </Dialog.Footer>
+          </Dialog.Content>
+        </Dialog.Positioner>
+      </Portal>
+    </Dialog.Root>
+  );
+}
+
+export default ThreadDeleteDialog;
